webapi/webapi-rawsockets-w3c-tests/rawsockets/UDPSocket_method.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset='utf-8'>
<title>Raw Socket Test: UDPSocket method</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/raw-sockets/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
// These cases still under designed status.
test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.close();
  assert_equals(socket.readyState, "closed", "The value of socket.readyState");
}, "Check if the value of readyState is closed after calling close method");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.joinMulticast("239.255.255.250");
  assert_unreached("TODO debug");
}, "Check if socket.joinMulticast can add the given address to join the multicast group");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.joinMulticast("239.255.255.250");
  socket.leaveMulticast("239.255.255.250");
  assert_unreached("TODO debug");
}, "Check if socket.leaveMulticast can leave the given address in the multicast group");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.suspend();
  assert_unreached("TODO debug");
}, "Check if socket.suspend can pause reading incoming UDP data");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.suspend();
  socket.resume();
  assert_unreached("TODO debug");
}, "Check if socket.resume can resume reading incoming UDP data");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  socket.joinMulticast("239.255.255.250");
  socket.send("test");
  assert_unreached("TODO debug");
}, "Check if socket.send can send data to the remoteAddress and remotePort");

</script>
